+++ lib/app/shell/shared/shell_edit_hosts.dart
@@ -1,4 +1,4 @@
-part of 'life_os_dashboard.dart';
+part of '../app_shell.dart';
 
 mixin ShellEditHosts on ShellCoreLogic {
   void openNewTaskForPastDate() {